auth-service/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
passlib[bcrypt]>=1.7
bcrypt==4.0.1
pydantic[email]>=2.6
httpx>=0.27
python-jose[cryptography]>=3.3
aio-pika>=9.0

# Test
pytest>=8.0
pytest-asyncio>=0.23
prometheus-fastapi-instrumentator>=6.1
